Resolved size_t issues

+++ ESP_ATMod/ESP_ATMod.h
@@ -68,7 +68,7 @@ extern uint16_t inputBufferCnt;				  // Number of bytes in inputBuffer
 extern uint8_t fingerprint[20];				  // SHA-1 certificate fingerprint for TLS connections
 extern bool fingerprintValid;
 extern BearSSL::X509List CAcert; // CA certificate for TLS validation
-extern int maximumCertificates;	 // Maximum amount of certificates to load
+extern size_t maximumCertificates;	 // Maximum amount of certificates to load
 
 extern char *PemCertificate;		 // Buffer for loading a certificate
 extern uint16_t PemCertificatePos;	 // Position in buffer while loading
